General cleanup of converts, moved to adapter
