+++ Etabs_Adapter/Etabs_Adapter.csproj
@@ -335,6 +335,10 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AdapterActions\Exectute.cs" />
+    <Compile Include="Convert\ToBHoM\Vector.cs" />
+    <Compile Include="Convert\ToCSI\BarRelease.cs" />
+    <Compile Include="Convert\ToCSI\Constraint6DOF.cs" />
+    <Compile Include="Convert\ToCSI\Vector.cs" />
     <Compile Include="CRUD\Create\Loadcase.cs" />
     <Compile Include="CRUD\Create\Diaphragm.cs" />
     <Compile Include="CRUD\Create\SurfaceProperty.cs" />
